Lesson 2 Complete
Lesson 2 - AutoLayout and Buttons
In lesson 2 we added 2 buttons: "Record" and "Stop Recording" to Main.storyboard, as well as a label "Tap to Record." The two buttons are linked to IBActions that print a string to the console, signaling that the button was pressed. IBAction recordAudio(_:) also modifies the text of the label "Tap to Record" to "Recording in Progress" which is linked to IBOutlet "recordingLabel." Each of the buttons and the label have constraints applied so that the view stays constant on both an iPhone as well as an iPad in both portrait and landscape orientations.