Lesson 3 Complete
Lesson 3 - ViewController and Multiple Views
In lesson 3 we added images to the two buttons found in the initial view in Main.storyboard. In this lesson we also added IBOutlets to the two buttons and used those outlets to enable and disable the record and stop recording buttons depending on which button was pressed. Finally we added a UINavigationController, embedded our initial view inside this navigation controller and added a second blank view controller to the application that appears when the stop recording button is pressed.