Lesson 4 Complete
Lesson 4 - Delegation and Recording
In lesson 4 we renamed the default ViewController.swift file that we had been working on to RecordAudioViewController.swift. This paved the way for a second view controller swift file to be added to the project called PlaySoundsViewController.swift. Second, in RecordAudioViewController we imported the AVFoundation framework and implemented the ability to record audio and pass that audio to the PlaySoundsViewController. The functionality to playback audio has not yet been implemented.