Lesson 5 Complete
Lesson 5 - Playback and Audio Effects
In lesson 5 we added seven buttons to the Play Sounds View Controller. Six of the buttons are used for audio playback with different audio effects, and the seventh is for stopping audio playback. The audio effect buttons are embedded in stack views that handle layout without the need for adding additional constraints. Finally we added audio playback functionality through the use of AV Audio Engine and an extension provided by Udacity.